+++ androidApp/src/main/kotlin/com/murzify/bambuddyspool/MainActivity.kt
@@ -6,14 +6,26 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import com.arkivanov.decompose.DefaultComponentContext
+import com.arkivanov.essenty.lifecycle.LifecycleRegistry
+import com.murzify.bambuddyspool.app.App
+import com.murzify.bambuddyspool.app.bootstrap.createRootGraph
+import com.murzify.bambuddyspool.core.platform.mockPlatformServices
 
 class MainActivity : ComponentActivity() {
+    private val root by lazy {
+        createRootGraph(
+            componentContext = DefaultComponentContext(LifecycleRegistry()),
+            platformServices = mockPlatformServices(),
+        ).rootComponent
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         enableEdgeToEdge()
         super.onCreate(savedInstanceState)
 
         setContent {
-            App()
+            App(root)
         }
     }
 }
@@ -21,5 +33,9 @@ class MainActivity : ComponentActivity() {
 @Preview
 @Composable
 fun AppAndroidPreview() {
-    App()
+    val root = createRootGraph(
+        componentContext = DefaultComponentContext(LifecycleRegistry()),
+        platformServices = mockPlatformServices(),
+    ).rootComponent
+    App(root)
 }