+++ shared/src/commonMain/kotlin/com/murzify/bambuddyspool/core/application/Scopes.kt
@@ -1,0 +1,11 @@
+package com.murzify.bambuddyspool.core.application
+
+import dev.zacsweers.metro.Scope
+
+@Scope
+@Target(AnnotationTarget.CLASS)
+annotation class ApplicationScope
+
+@Scope
+@Target(AnnotationTarget.CLASS)
+annotation class ComponentScope